+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Improvements
+
+- [#726](https://github.com/cosmos/iavl/pull/726) Make `KVPair` and `ChangeSet` serializable with protobuf.
+
 ### Breaking Changes
 
 - [#646](https://github.com/cosmos/iavl/pull/646) Remove the `orphans` from the storage